+++ packages/system/src/utils/general.ts
@@ -184,7 +184,7 @@ export const rendererPointToPoint = ({ x, y }: XYPosition, [tx, ty, tScale]: Tra
  */
 function parsePadding(padding: PaddingWithUnit, viewport: number): number {
   if (typeof padding === 'number') {
-    return Math.floor(viewport - viewport / (1 + padding));
+    return Math.floor((viewport - viewport / (1 + padding)) * 0.5);
   }
 
   if (typeof padding === 'string' && padding.endsWith('px')) {
@@ -398,4 +398,23 @@ export function areSetsEqual(a: Set<string>, b: Set<string>) {
   }
 
   return true;
+}
+
+/**
+ * Polyfill for Promise.withResolvers until we can use it in all browsers
+ * @internal
+ */
+export function withResolvers<T>(): {
+  promise: Promise<T>;
+  resolve: (value: T | PromiseLike<T>) => void;
+  reject: (reason?: unknown) => void;
+} {
+  let resolve!: (value: T | PromiseLike<T>) => void;
+  let reject!: (reason?: unknown) => void;
+
+  const promise = new Promise<T>((res, rej) => {
+    resolve = res;
+    reject = rej;
+  });
+  return { promise, resolve, reject };
 }